Add support for signed data in Support Library
* Cascading compiler needs to fill the MceScheduler agent data to
enable the support for signed data. This patch sets the proper
signedness (isIfmSigned and isOfmSigned)information for that purpose.

* Zero point info associated with IFM/OFM streamer in case of FCAF
data-format wasn't being set properly in the case of signed input/
output because prior to this patch zero point value was defined as an
uint8 which could result into erroneous values if it's value is
between -128 and 0. Therefore, the type of zero point for IFM/OFM
streamer was bumped to int16.

+++ driver/support_library/command_stream/include/ethosn_command_stream/cascading/CommandStream.hpp
@@ -45,7 +45,7 @@ enum class FmsDataType : uint8_t
 struct FcafInfo
 {
     /// Zero point info needed for FCAF
-    uint8_t zeroPoint;
+    int16_t zeroPoint;
     /// Signed activation info needed for FCAF
     bool signedActivation;
 };
